+++ src/uci.h
@@ -11,7 +11,7 @@ void loop();
 class Info {
 public:
     void clear();
-    void update(const Position& pos, int depth, int score, int nodes, std::vector<move_t>& pv,
+    void update(const Position& pos, int depth, int score, uint64_t nodes, std::vector<move_t>& pv,
                 bool partial = false);
     void print_bestmove(const Position& pos) const;
